# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "University Management API",
        "description": "REST API for the Laravel 12 University Management System. All protected endpoints require a Sanctum Bearer token obtained from POST /api/auth/login.",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/v1/courses": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "List all courses",
                "operationId": "a4db43e450f4b2dd50c56d0e061ee2a7",
                "responses": {
                    "200": {
                        "description": "Array of courses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Course"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Courses"
                ],
                "summary": "Create a new course",
                "operationId": "189ef8d6de117833128c35345606a73b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "course_code",
                                    "title",
                                    "credit_hours"
                                ],
                                "properties": {
                                    "course_code": {
                                        "type": "string",
                                        "example": "CS101"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Introduction to Computer Science"
                                    },
                                    "credit_hours": {
                                        "type": "integer",
                                        "example": 3,
                                        "maximum": 6,
                                        "minimum": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Course created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Course"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/courses/{id}": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get a course by ID",
                "operationId": "1ddb9f5c4468a1a508891a2d1851943e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Course with relationships",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Course"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Courses"
                ],
                "summary": "Update a course",
                "operationId": "bda600cc445c589e56806c0eefdf4ba8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "course_code": {
                                        "type": "string",
                                        "example": "CS101B"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Intro to CS (Updated)"
                                    },
                                    "credit_hours": {
                                        "type": "integer",
                                        "example": 4,
                                        "maximum": 6,
                                        "minimum": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "updated": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Courses"
                ],
                "summary": "Delete a course",
                "operationId": "fe893d7cce54920fdc51ef62d9d369c3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "deleted": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/courses/{id}/available-students": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "List students not yet enrolled in the course",
                "operationId": "9a8f923cbb249976ce5bae1e8f99dba7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of available students"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/courses/{id}/assigned-students": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "List students enrolled in the course",
                "operationId": "06fd52ebb04decced4a7a950282a4798",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of assigned students"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/courses/{id}/assign-students": {
            "post": {
                "tags": [
                    "Courses"
                ],
                "summary": "Enroll multiple students in a course",
                "operationId": "9ce846cf0f2565f2a5f2451ca3f8d7b5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "student_ids",
                                    "semester"
                                ],
                                "properties": {
                                    "student_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2,
                                            3
                                        ]
                                    },
                                    "semester": {
                                        "type": "string",
                                        "example": "2024-S1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Assignment result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "assigned_count": {
                                            "type": "integer"
                                        },
                                        "conflict_count": {
                                            "type": "integer"
                                        },
                                        "conflict_students": {
                                            "type": "array",
                                            "items": {
                                                "type": "integer"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/v1/courses/{id}/available-instructors": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "List instructors not yet assigned to the course",
                "operationId": "9383cac4ebb2972d96899b74cbb89316",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of available instructors"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/courses/{id}/assigned-instructors": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "List instructors assigned to the course",
                "operationId": "c90bf6d5b5d27a72b794c21a0682c36a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of assigned instructors"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Course": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "course_code": {
                        "type": "string",
                        "example": "CS101"
                    },
                    "title": {
                        "type": "string",
                        "example": "Introduction to Computer Science"
                    },
                    "credit_hours": {
                        "type": "integer",
                        "example": 3
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication. Obtain a token via POST /api/auth/login then enter it here (without the \"Bearer \" prefix — Swagger UI adds that automatically).",
                "scheme": "bearer",
                "bearerFormat": "token"
            }
        }
    },
...
}